Move and integrate lp_copy_packages and utils
